10/19 Submission 1.0
Added "@starbuckslover1989" to reflect final hi-fi version. Still plan on updating img since the carousel circles moved to top instead of bottom in hi-fi (the carousel is not functional in the code, I originally exported the image with the carousel circles already on it). And I plan on adding "260 tried it" under the skill/time estimate